refactor(core): move the field-configuration hook into FieldComponentBase
